Create Key Levels Spaceman + 6+ Candles
PineScript Code specifically for the use of TradingView. This code combined SpacemanBTC Chart Levels which uses time and candle placement for figuring out some technical analysis. 

In combination with this code, it correctly combines an indicator built by a friend to show a change in candle color/overlayed candle color when 6+ candles are in the same direction. 

Example: 6 red candles, 7th candle is displayed as a different color, set for the specific red candles that show the higher continuation of the current trend until the color flips to, for this example, green. 
Vice versa this example for 6 green candles, so from green to a special color, and then to red when the continuation of the trend stops.